update tarea6
change data type of matrix and create a makefile

+++ tarea6/ClaseRMI.java
@@ -8,8 +8,8 @@ public class ClaseRMI extends UnicastRemoteObject implements InterfaceRMI {
     super();
   }
 
-  public int[][] multiplica_matrices(int[][] A, int[][] B) throws RemoteException {
-    int[][] C = new int[N / 2][N / 2];
+  public float[][] multiplica_matrices(float[][] A, float[][] B) throws RemoteException {
+    float[][] C = new float[N / 2][N / 2];
     for (int i = 0; i < N / 2; i++)
         for (int j = 0; j < N / 2; j++)
             for (int k = 0; k < N; k++)
