+++ tarea6/InterfaceRMI.java
@@ -2,5 +2,5 @@ import java.rmi.Remote;
 import java.rmi.RemoteException;
 
 public interface InterfaceRMI extends Remote {
-  public int[][] multiplica_matrices(int[][] A, int[][] B) throws RemoteException;
+  public float[][] multiplica_matrices(float[][] A, float[][] B) throws RemoteException;
 }
